Update Node engine specified in package.json
Hi there! Thanks for this package :)

[See here](https://github.com/github/hubot/issues/1009#issuecomment-264596324) for a bit more context on this PR.

This updates the template `package.json` to specify a more recent version of Node, one which supports ES6 by default. I'm writing a bot that uses ES6 scripts so I noticed this issue when I tried to deploy to Heroku. Everything ran fine on my laptop, because I have a recent version of Node which supports ES6, but Heroku was using the version specified in `package.json`, which doesn't support ES6 by default.

This isn't a permanent solution, because it will become out-of-date again before too long, but I hope this helps. Please let me know if you have any questions.